Changed the itemDataloader to fit with the new database Changed the way that items are loaded and stored in the database. Instead of being stored in one large list they are now stored in three separate lists one for buffs, debuffs and vanityItems.

+++ KahIT/app/src/main/java/com/god/kahit/model/Debuff.java
@@ -2,7 +2,7 @@ package com.god.kahit.model;
 
 public class Debuff extends Modifier {
     private Boolean autoAnswer;
-    public Debuff(int price, String name, int scoreMultiplier, int timeHeadstart, boolean autoAnswer){
+    public Debuff(int price, String name, double scoreMultiplier, int timeHeadstart, boolean autoAnswer){
         super(price, name, scoreMultiplier, timeHeadstart);
         this.autoAnswer = autoAnswer;
     }
